+++ AzureBatchQueue/MessageBatch.cs
@@ -1,5 +1,4 @@
 ﻿using Azure.Storage.Queues;
-using Azure.Storage.Queues.Models;
 using Newtonsoft.Json;
 
 namespace AzureBatchQueue;
@@ -48,52 +47,3 @@ public class MessageBatch<T>
 }
 
 public record MessageBatchOptions(string MessageId, string PopReceipt, TimeSpan FlushPeriod);
-
-public class BatchQueue<T>
-{
-    private readonly QueueClient queue;
-    private readonly TimeSpan flushPeriod;
-
-    public BatchQueue(QueueClient queue, TimeSpan flushPeriod)
-    {
-        this.queue = queue;
-        this.flushPeriod = flushPeriod;
-    }
-
-    public async Task SendBatch(IEnumerable<T> items)
-    {
-        await queue.SendMessageAsync(MessageBatch<T>.Serialize(items));
-    }
-
-    public async Task<BatchItem<T>[]> ReceiveBatch()
-    {
-        var msg = await queue.ReceiveMessageAsync();
-        var items = Deserialize<T>(msg.Value);
-        var batchOptions = new MessageBatchOptions(msg.Value.MessageId, msg.Value.PopReceipt, flushPeriod);
-
-        var batch = new MessageBatch<T>(queue, items, batchOptions);
-
-        return batch.Items();
-    }
-
-    private static IEnumerable<T> Deserialize<T>(QueueMessage value) => value.Body.ToObjectFromJson<T[]>();
-}
-
-public class BatchItem<T>
-{
-    public BatchItem(Guid id, MessageBatch<T> batch, T item)
-    {
-        Id = id;
-        Batch = batch;
-        Item = item;
-    }
-
-    public Guid Id { get; }
-    private MessageBatch<T> Batch { get; }
-    public T Item;
-
-    public async Task Complete()
-    {
-       await Batch.Complete(Id);
-    }
-}
